MOVEONLY: Move spending RPCs to spend.cpp
